Created Custom FormField for required! Cleaned up stories
